pkg/apis/monitoring/v1: separate tests into each own file
